google: Remove sort after find in gen_initramfs_list.sh
This patch removes a sort command after a find command in
gen_initramfs_list.sh, used to process a directory of files. The sort
is meant to order the files in the cpio archive. However, it produces
the following kind of output:

    /lib/modules/4.13/...
    /lib/modules

This resuls in a cpio archive with the /lib/modules entry after all of
its children node entries. This in turn breaks the kernel's cpio unpack
to rootfs implementation where none of the /lib/modules/... nodes are
created because the parent node does not exists when they are visited.

Removing the sort fixes this issue because the find command by default
visits depth-first and prints parent nodes before children.

Change-Id: I0c504c8103287554252de5df83384267511e537e