SWDEV-191980:drm/amd/display: Expose HDR output metadata for supported connectors
[Why]
For userspace to send static HDR metadata to the display we need to
attach the property on the connector and send it to DC.

[How]
The property is attached to HDMI and DP connectors. Since the metadata
isn't actually available when creating the connector this isn't a
property we can dynamically support based on the extension block
being available or not.

When the HDR metadata is changed a modeset will be forced for now.
We need to switch from 8bpc to 10bpc in most cases anyway, so this isn't
completely unnecessary. The requirement could later be reduced to just
entering and exiting HDR or switching bpc.

Change-Id: Ib9a47dbc16800b7cecee71d3a7106139182d8a15
Signed-off-by: Nicholas Kazlauskas <nicholas.kazlauskas@amd.com>
Signed-off-by: Emily Deng <Emily.Deng@amd.com>