google: Mount boot and kroot filesystems above all others
This patch modifies the Stadia overlayfs script to mount the boot and
kroot filesystems above all others. This ensures that the booted kernel
and its associated root filesystem content (modules and firmwares) take
precedence over files provided by other filesystems included in the root
overlayfs.

Testing:
Run mount in booted VM and verified that bootfs and krootfs are "on the
left" ("above") the other root overlayfs filesystems.

Run "perf" and check that it works. The krootfs provides an "perf" that
overrides the distro's "perf".

Bug: 135691167
Change-Id: I9947f5832957b9af161f77a0c3b34e4f7e1d7d04