SWDEV-197284 - drm/amdgpu: Only use the peek function in productor side is not correct
For spsc queue, use peek function would cause error in productor side.
As for the last element, when the push job is occurring during pop job, the peek function
will not be updated in time, and it will return NULL.

So use queue count for double confirming the job queue is empty.

Signed-off-by: Emily Deng <Emily.Deng@amd.com>