SWDEV-216491 - [Gibraltar][one-VF_mode vbios]RGP test failed
drm/amd/powerply: wait before query clock

Just one workaround for RGP test app problem.
RGP will directly query clock after setting it in 1ms
which is not correct because when SMU response register is 1
it only means SMU has done the setting which not indicate the
actual clock archieve the target. Especially for mem clock,
it archieve the target will be later than gfx clock.
The reason why legacy SIMPERF will not encounter this issue
is that the handshake between vf and pf to query clock will
cost about 10ms which avoid the problem.

Signed-off-by: Yintian Tao <yttao@amd.com>
Change-Id: I80f6ea165950f46927c2c24d6db0b67f44feaf9e