SWDEV-226000 - [Gibraltar][navi12 weekly][20200228] FLR guard is not off after wait for 60 sec
If, on host side, FLR event guard is overflow, then host
would not response to reset request in 60 seconds. Guest
should keep polling for more than 60 seconds in this case.

On my server, msleep(10) takes about 20ms, so I change the
polling period to be 6000ms, so that the effective period
would double, 12000ms. In 5 or 6 rounds, it would get response
from host.

However, on different machines, msleep(10) would take shorter
time or longer time. In most accurate case, it would take 10ms,
so the effective period is 6000ms. In this case it would need 10
or 11 rounds to get response from host.

Change-Id: I1472cbe491b569174ed6c485930a5ede0e5d7dd5
Signed-off-by: jianzh <Jiange.Zhao@amd.com>