Fix AppArmor/OverlayFS issue
The when and why the AppArmor/OverlayFS issue happens can been seen at
go/yeti-apparmor-overlayfs. It is a known issue in the Linux kernel and
requires kernel code changes for it to be fixed.

This change adds a helper function in namei.c for finding the real
vfsmount for a provided path, and AppArmor make use of it in path.c to
find the real path name of a disconnected directory.

One caveat of this is if the real path is bind-mounted (in our case
/mnt/developer is bind mount to /mnt/localssd/developer), then the new
helper function will return the vfsmount of /mnt/localssd/developer and
thus cause AppArmor to see that path. Fortunately for us, if we're
already allowing access to /mnt/developer, then allowing
/mnt/localssd/developer should not be a problem.

I don't know if this is good enough to go into the official Linux kernel
repo since the new helper function is not very general and has the
drawback mentioned above. But I believe it's good for us, and it does
fix the issue that has existed for a long time and no one has fixed it,
so I may contact the owner of AppArmor and ask them what they think
about this after fixing it for Stadia first.

Bug: 157263110
Fixes: 157263110
Test: Built a kernel and ran it on a devflow Gamelet. Verified that it
fixed the AppArmor issue we were facing.
Also pushed the kernel to a preprod gamerflow Gamelet and ran a game
succesfully.

Change-Id: I14ecb5d9041f22f743ccc4b617a91dcea6295573