google: Mount the EFI system partition at /boot
The EFI system partition is a VFAT filesystem which cannot be included
in overlay filesystems. This patch modifies the EFI system partition to
include the kernel and other boot files directly at the root the EFI
system partition, and modifies the overlayfs script to mount the EFI
system partition at /boot on top of the overlayfs root filesystem.

Bug: 153885920
Bug: 162757480
Change-Id: I9002831eb7647f2e8734d2e0c3d0df01b8991879