google: drm/amdgpu: Do not use exclusive fence for shared Bos.
For Vulkan, amdvlk populates bo_list with all the global memories even
if those are not used. This is due to the Vulkan design making it
difficult to determine the usage of a buffer in all cases such as if
vkGetBufferDeviceAddressKHR is used to get the device address of the
buffer and used directly. During scheduling in KMD it validates the
buffers in bo_list. Synchronization with the buffers that are
moved/evicted is done by adding shared fences to them.  Jobs that are
using those buffers are synchronized with these fences.

For shared buffers it uses exclusive fences, jobs are always
synchronized with exclusive fences. Because it adds all the buffers
(including shared) in each job, this makes the jobs to become
synchronized when a buffer is imported. Note that this would happen even
if the buffer is never used.

Since in Vulkan, buffers are explicitely synchronized by using
synchronization objects, the implicit synchronization of shared buffers
is not required. Temporarily remove the explicit synchronization of
shared buffers until we have a better solution for this.

Tested:
- Executed 3dmark with change and the gpu queues run asynchronously.

Google-Bug-Id: 186735161
Change-Id: I47275bfbd38840553b945f3822c6d0d4f872bb5e