TracerPid is pid_max for parent namespace tracer
For the Stadia debugger flow to work, the game process needs to be able
to detect when it is being traced by a process in its parent PID
namespace, which TracerPid in /proc/[pid]/status won't show. This change
will make TracerPid return pid_max instead of 0 for this case. This will
enable checks that look to see if they are being traced to return true,
even though they can't see the actual process.

Change-Id: I828da9bd59eb3b283ad06587c7dae442169d1064