google: memremap: Change P2PDMA caching mode to writecombine
The P2PDMA was set to UC in this commit since it was going to default (WB):
https://github.com/torvalds/linux/commit/a50d8d98a87f33efa07adfa20747e13a93839a4b

The commit also states that the future usecases might want to select a
different cache type as they might not want UC. This is true in our case,
since our GPU VRAM is remapped with writecombine cache mode during startup.

This will be later replaced by a generalize way to allow the caller to
specifiy the cache mode as a flag/parameter.

Change-Id: Ib49db5f524e220eb41af82a749b6daf1992c3f70