Add support for building NVIDIA UMD/KMD disk
Add support to the kokoro job script to generate
a disk that contains the UMD/KMD NVIDIA modules
and corresponding support files required for
instances that use a NVIDIA gpu.

The disk is generated from a preprocessed package
that contains NVIDIA kernel headers and UMD stored
in GCS.

To create an NVIDIA disk the script requires setting
the NVIDIA_DRIVER_VERSION env var. If the var is not
set it will generate a new kernel image with the AMD
modules.

The disk is not self contained and must be mounted
with the corresponding kernel disk.

Bug: 231133651
Change-Id: I8547c908bc68fe86877139da962008a7a93c3f9a